added staging env check

+++ backend/authentication_app/jobs/daily/delete_devicelist.py
@@ -1,3 +1,4 @@
+import os
 from django_extensions.management.jobs import DailyJob
 from authentication_app.models import DeviceList
 
@@ -6,10 +7,15 @@ class Job(DailyJob):
     help = "Delete Device List"
 
     def execute(self):
-        print("Deleting Device List...")
-        try:
-            DeviceList.objects.all().delete()
-            print("Deleted Device List")
-        except:
-            print("Error deleting Device List")
-            return
+        if ("KORANGLE_STAGING" in os.environ) and (
+            os.environ["KORANGLE_STAGING"] == "TRUE"
+        ):
+            print("Deleting Device List...")
+            try:
+                DeviceList.objects.all().delete()
+                print("Deleted Device List")
+            except:
+                print("Error deleting Device List")
+                return
+        else:
+            print("Please only run this job in the Staging Environment")
